requirements.txt
# NotebookLM Skill Dependencies
# These will be installed in the skill's local .venv

# NotebookLM API client (includes playwright for login)
notebooklm-py[browser]>=0.3.3

# SOCKS proxy support for httpx
httpx[socks]

# Environment management
python-dotenv==1.0.0
